# machinesaggregateservice/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CPU:
    properties:
      available:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      total:
        type: integer
    type: object
  model.Category:
    properties:
      gpu:
        $ref: '#/definitions/model.GPUDict'
    type: object
  model.GPU:
    properties:
      available_vram:
        type: integer
      avg_dlperf:
        type: number
      id:
        type: string
      name:
        type: string
      price:
        type: number
      total_vram:
        type: integer
      used_vram:
        type: integer
    type: object
  model.GPUDict:
    properties:
      id:
        type: string
      name:
        type: string
      price:
        type: number
      price_min:
        type: number
      total_sessions:
        type: integer
      total_vram:
        type: number
      type:
        type: string
    type: object
  model.PrePull:
    properties:
      id:
        type: string
      template_id:
        type: string
    type: object
  model.Session:
    properties:
      available_ram:
        type: integer
      avatar:
        type: string
      cpus:
        items:
          $ref: '#/definitions/model.CPU'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      gpus:
        items:
          $ref: '#/definitions/model.GPU'
        type: array
      load_speed:
        type: number
      ping:
        type: number
      prepull:
        items:
          $ref: '#/definitions/model.PrePull'
        type: array
      price_internet:
        type: number
      price_ram:
        type: number
      reliability:
        type: number
      score:
        type: number
      session_id:
        type: string
      storages:
        items:
          $ref: '#/definitions/model.Storage'
        type: array
      total_price:
        type: number
      total_ram:
        type: integer
      upload_speed:
        type: number
      used_ram:
        type: integer
      user_id:
        type: string
    type: object
  model.SessionList:
    properties:
      max_ram:
        type: integer
      max_storage:
        type: integer
      min_ram:
        type: integer
      min_storage:
        type: integer
      sessions:
        items:
          $ref: '#/definitions/model.Session'
        type: array
    type: object
  model.SessionResponse:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      session:
        $ref: '#/definitions/model.Session'
    type: object
  model.Storage:
    properties:
      available:
        type: integer
      bandwidth:
        type: number
      id:
        type: string
      name:
        type: string
      price:
        type: number
      total:
        type: integer
      type:
        type: string
      used:
        type: integer
    type: object
  model.Tariff:
    properties:
      gpu:
        $ref: '#/definitions/model.GPUDict'
      id:
        type: string
      price:
        type: number
      sessions:
        items:
          type: string
        type: array
      total_sessions:
        type: integer
      type:
        type: string
    type: object
  response.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is an API for MachinesAggregation service
  title: Swagger MachinesAggregation Service Api
  version: "1.0"
paths:
  /exchange/category/gpu:
    get:
      consumes:
      - application/json
      description: Create feeedback for rental.
      parameters:
      - description: vram_from
        in: query
        name: q
        type: string
      - description: vram_to
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success receiving gpu categories
          schema:
            items:
              $ref: '#/definitions/model.GPUDict'
            type: array
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/response.HTTPError'
        "500":
          description: Something went wrong
          schema:
            $ref: '#/definitions/response.HTTPError'
      summary: List of gpu categoties
  /exchange/category/gpu/{gpu_id}/session:
    get:
      consumes:
      - application/json
      description: This endpoint returns sessions by ID of gpulist.
      parameters:
      - description: gpuID
        in: path
        name: gpu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success receiving sessions
          schema:
            $ref: '#/definitions/model.SessionList'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/response.HTTPError'
        "500":
          description: Something went wrong
          schema:
            $ref: '#/definitions/response.HTTPError'
      summary: Receive list sessions by gpuIDlist endpoint
  /exchange/category/tariff:
    get:
      consumes:
      - application/json
      description: This endpoint returns information about tariffs and information
        about available sessions.
      produces:
      - application/json
      responses:
        "200":
          description: Success receiving tariffs
          schema:
            items:
              $ref: '#/definitions/model.Tariff'
            type: array
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/response.HTTPError'
        "500":
          description: Something went wrong
          schema:
            $ref: '#/definitions/response.HTTPError'
      summary: List of tariffs endpoint
  /exchange/category/tariff/{tariff_id}/session:
    get:
      consumes:
      - application/json
      description: This endpoint returns sessions by ID of tariff.
      parameters:
      - description: tariffID
        in: path
        name: tariff_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success receiving sessions
          schema:
            $ref: '#/definitions/model.SessionList'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/response.HTTPError'
        "500":
          description: Something went wrong
          schema:
            $ref: '#/definitions/response.HTTPError'
      summary: Receive list sessions by tariffID endpoint
  /session/{session_id}:
    get:
      consumes:
      - application/json
      description: This endpoint returns session by ID.
      parameters:
      - description: sessionID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success receiving session
          schema:
            $ref: '#/definitions/model.SessionResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/response.HTTPError'
        "500":
          description: Something went wrong
          schema:
            $ref: '#/definitions/response.HTTPError'
      summary: Receive session endpoint
swagger: "2.0"